feat: enable replacing the lms name that aispot will be given
This lets us handle devstack, where the lms will be visible from the
aispot container at "lms" but not at "localhost". Excatly what that
name is depends on docker setup.

Production is not expected to use this setting since the aside
should be able to generate a perfectly usable URL in that case.