feat: exception sponge in summaryhook
We don't want the summary aside to every be able to block rendering
course content. That means the check for whether the aside should
apply and the actual aside html injection need to be wrapped in
a statement which soaks all exceptions.

This does not wrap the summary content handler endpoint - if that
breaks it will prevent summarization but will not block the learner
and errors will be visible as normal endpoint errors.